requirements.txt
# Core data science
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15

# Network analysis
networkx>=3.1
pyvis>=0.3.2

# Streamlit demo
streamlit>=1.30

# O*NET API (supplementary)
requests>=2.31
python-dotenv>=1.0

# Jupyter
jupyter>=1.0
ipykernel>=6.25

# Utilities
openpyxl>=3.1        # Excel file reading
tqdm>=4.65           # Progress bars
